// src/integrations/supabase/types.ts
export type Database = {
  public: {
    Tables: {
// ...
      profiles: {
        Row: {
          county: string
          created_at: string | null
          full_name: string
          id: string
          is_verified: boolean | null
          location_coordinates: unknown | null
          phone_number: string
          rating: number | null
          sub_county: string | null
          total_transactions: number | null
          updated_at: string | null
          user_role: Database["public"]["Enums"]["user_role"]
        }
        Insert: {
          county: string
          created_at?: string | null
          full_name: string
          id: string
          is_verified?: boolean | null
          location_coordinates?: unknown | null
          phone_number: string
          rating?: number | null
          sub_county?: string | null
          total_transactions?: number | null
          updated_at?: string | null
          user_role: Database["public"]["Enums"]["user_role"]
        }
        Update: {
          county?: string
          created_at?: string | null
          full_name?: string
          id?: string
          is_verified?: boolean | null
          location_coordinates?: unknown | null
          phone_number?: string
          rating?: number | null
          sub_county?: string | null
          total_transactions?: number | null
          updated_at?: string | null
          user_role?: Database["public"]["Enums"]["user_role"]
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}

type DefaultSchema = Database[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends { schema: keyof Database }
  ? Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never